Add a model for general comments made on a review.
Reviewboard currently support 3 types of comments: Diff commment, File
attachment comments, Screenshot comments. This is a new type of comments:
General comments.

A general comment on a review request is used when a comment is not tied to
specific lines of code or a special file attachment. Examples include
suggestions for testing or pointing out errors in the change description.

This review request is for GeneralComment backend model.

Reviewed at https://reviews.reviewboard.org/r/6375/