Fixes tabs appearing in the subject line of emails.
Python 2.7 has a problem that causes tabs to be used to break up long headers
instead of spaces. When Review Board sends an email with a Subject longer than
62 characters, it inserts a tab to seperate the text into mutiple lines.

Now instead of the Subject header being seperatated by tabs it is seperated by
spaces.

This was originally pushed as c31bdf5, but we had to revert it because it broke
for e-mail subjects containing unicode. This updated version fixes that.

Testing Done:
Ran unit tests.

Bugs closed: 3613, 3707

Reviewed at https://reviews.reviewboard.org/r/6647/