Update email notification logic to honor should_send_own_update flag
If a user was explicitly listed in a review request along with also being a
member of a group assigned to a request, the should_send_own_update setting in
the users profile would not be honored because the check was happening before
we added explicit users to the to_field list. This also fixes the issue where
we were trying to remove email address that might not actually exist (a blank
review request).

Testing Done:
Had a user create a review request that lists me as an explicit reviewer along
with a group that I'm also a member of; commented on the issue and I never
recieved an email notification.

Bugs closed: 3681, 3684

Reviewed at https://reviews.reviewboard.org/r/6663/