Add UI for creating general comments.
Previous changes introduced backend support for general comments on a review
request. A general comment on a review request is used when a comment is not
tied to specific lines of code or a special file attachment, and an issue is
opened. Examples include suggestions for testing or pointing out errors in the
change description.

This change adds UI for creating these directly on the review request.
The action bar now includes an "Add Comment" button that pops up a comment
dialog for creating a general comment.

General comments are shown in the review dialog, before any other comments.
They can be edited and the "Open an issue" checkbox can be toggled.

Reviewed at https://reviews.reviewboard.org/r/6506/