Add FileAttachment UUIDs, and a view for redirecting to the attachment.
Part of a set of changes to add support for dragging and dropping images into
the markdown editor.

This adds a 'uuid' field to FileAttachment that can be used to identify a
particular file attachment with a non-guessable identifier.

A view has been added to redirect to the path of a user-owned file attachment
identified by a UUID.

Reviewed at https://reviews.reviewboard.org/r/6616/