Limit e-mail recipients when adding reviewers.
Occasionally, the only change to a review request will be to add additional
reviewers. This is especially common when a review request is first sent to a
generic group, and then people in that group make suggestions about who might
be more appropriate to review the code in question.

This change makes it so that when the only change to a review request is to
change the list of reviewers (people or groups), the e-mail that is sent is
sent only to the author of the review request (assuming that they have "own
actions" e-mails turned on), people who have starred the review request, and
the newly-added reviewers. This was requested by a user who has some very
high-volume e-mail lists, in an effort to cut down on the amount of traffic.

In the case where the change was to remove reviewers, the e-mail is sent only
to the author. In the case that the author has "own actions" e-mails turned
off, no e-mail will be sent.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/6683/