Remove old VMwarePerforceTool implementation.
For most of Review Board's history, we've shipped a customization of the
Perforce SCMTool tailored to the needs at a specific company, under the guise
of it providing an example. With the rich ecosystem of SCMTool implementations
we have now, that example is much less useful, and it's pretty weird to have it
in the core. This change removes it.

Testing done:
Ran unit tests.