Sandbox reviewboard.admin.widgets.Widget
Extensions that create a Widget subclass (using the AdminWidgetHook) can throw
exceptions inside Reviewboard. The parts of Reviewboard that call those methods
have been sandboxed.

Now when a Widget subclass from an extension throws an exception; Reviewboard
logs the errors with enough information to find which method in the Widget
subclass threw the exception.

Testing Done:
Unit tests have been written to make sure that functions from a Widget subclass
have been called, and when an exception is thrown it gets logged.

The tests fail without the sanboxing, and succeed with it.

Admin UI displays the widgets in the correct column.
When the "X" in the corner is clicked it closes the widget, but does refresh the page.

Reviewed at https://reviews.reviewboard.org/r/6650/