Focus Username field on login
Originally, the cursor focus would not be on the Username field in the login
page. This resulted in users have to press tab multiple times or click on the
Username field before filling in the username. With the fix, autofocus is added
to the Username field so that the Username field is focused by default.

Testing Done:
- Reloaded the login page, and made sure that cursor / focus is on Username field.
- Ensured login by email is still functional.
- Attempted login
- Attempted to change focus

Bugs closed: 3643

Reviewed at https://reviews.reviewboard.org/r/6731/