"Edit" button appears next to image titles that the user doesn't have permission to edit
The Edit button on image captions was shown to everyone, as the check was not
done on the front end. Checking with the model for ReviewRequestEditor was done
to ensure edit permissions were allowed.

Testing Done:
Testing was done manually with admin, users (people who created the issue, part
of the group/issued to, or not part of the group) and visitors.

Bugs closed: 3606

Reviewed at https://reviews.reviewboard.org/r/6747/