Show an error message when trying to compare different file attachment types.
The file attachment revisioning stuff doesn't actually put any restrictions on
the file type that's used to replace an attachment. This was fine when just
displaying individual revisions, but the comparison views would break in fun
and interesting ways.

This change fixes up our existing attachment UIs to show an error message in
the case where the comparison won't work. This currently has to be done
individually for each different Review UI, which isn't ideal. I'll make sure
that we get an item added to our backlog to try to clean that up in the future.

Testing done:
- Set up a file attachment with many different revisions, each of a different
  type. Verified that I saw the correct thing in each of the different Review
  UIs (which get selected based on the "orig" version being shown).
- Ran unit tests.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/6741/