Pre-fill the file caption when updating.
When updating a file attachment, it's likely that the user won't want to change
the caption. In this case, presenting a blank caption field will encourage them
to fill something in, resulting in different revisions having different
captions. This change makes it so that the field in the upload dialog is
pre-filled with the existing caption when updating.

Testing done:
- Clicked "Add File" and saw the caption field was empty.
- Clicked "Update" and saw the caption field was pre-filled with the old
  caption.

Reviewed at https://reviews.reviewboard.org/r/6755/