Fix the URL format for GitWeb raw file URL masks.
GitWeb-backed Git repositories had to be configured with a particular
URL, but the format we provided for the URL was incorrect. The hb=
parameter was given, but it should have been h=.

This updates the docs and the example to use the correct parameter.

Reviewed at https://reviews.reviewboard.org/r/6775/