Add a dummy review UI.
This change adds a new dummy review UI for file types that otherwise don't have
one. The purpose here is that if there's an attachment with multiple revisions,
and one of those revisions doesn't have a review UI, it would be possible to
get a 404 when sliding around through the revision slider.

The dummy UI just shows the revision header, selector, and caption/revision
label. It doesn't have anything else, or any commenting functionality. It's
highly unlikely that this will ever actually be encountered, but it's here just
in case.

While doing this, I fixed a bug that slipped into the image review UI captions,
and started combining some of the common parts of the Review UI headers.

Testing done:
- Verified that the dummy UI looked correct for both single and diff views.
- Checked that the thumbnail did not link to the dummy review UI.
- Ran python unit tests.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/6762/