Fix rendering change descriptions without raw insert/delete counts.
When rendering a change entry for a diff, the code attempts to get the
total raw insert/delete line counts across all FileDiffs. If there are
no FileDiffs for some reason, then this would end up crashing.

We're now more careful to not assume these counts exist, and to use 0 as
the value in this case.

Reviewed at https://reviews.reviewboard.org/r/6778/