Fix updating the changenum field with a PUT call to ReviewRequest
Making a PUT request to the ReviewRequest webapi with an updated changenum
field such as:

    `curl http://reviewboard.example.com/api/review-requests/8/ -X PUT -d changenum=1234`

for a perforce review request would not save the commit_id properly. This lead
to two modifications made to the PUT command regarding changenum. The changenum
is converted using six.text_type to make sure that it is a string, and the save
logic has been updated to make sure that the review request object is saved
with the correct update_fields.

Testing Done:
Two new test cases were added:
- Tests the initialization of ReviewRequest with changenum, and checks if the
  instance has the proper changenum and an empty commit_id.
- Tests the PUT call with changenum, and ensures the payload and the
  ReviewRequest saved have their changenum and commit_id changed to match up.
This includes the implementation for TestTool.get_changeset was made, for
scmtools that support pending changesets.

Bugs closed: 3685

Reviewed at https://reviews.reviewboard.org/r/6774/