Change Descriptions are no longer added when initially publishing.
This is a fix for the bug I introduced in the fix to 3465.

With this fix:
 - Change descriptions are no longer published on initial review request
   publishings.
 - Discarding no longer modifies the fields in the change description
   (aside from status, which it always did)

The refactored code pretty much changes the state prior to the shipped
fix of 3465, and extracts the "copying over from draft to review
request" part to a separate function. That function is then called if a
private unpublished review request is discarded.

Testing Done:
- Ran existing unit tests & passed
- Modified previous unit tests that I wrote to reflect the "copying
  over" part and not the change description. Passed.
- Added a test to make sure change descriptions are no longer published
  on initial review request publishings.
- Also manual tests of the above!

Bugs closed: 3465

Reviewed at https://reviews.reviewboard.org/r/6767/