Add a new http digest authentication backend
This change adds an authentication backend to allow Review Board to use http
digest files for authentication. This backend includes a settings form so that
you can specify the location of the file and the realm used for digest auth.

Testing Done:
Have tested it with various digest password files in the format of
username:realm:passwordhash where the password hash is comprised of the MD5
hash of "user:realm:pass". Files can be created using htdigest.

Reviewed at https://reviews.reviewboard.org/r/6812/