Don't mark incoming rich text in replies as plain text.
When replying to a review and turning off Markdown, we'd make an API
request to the server that, along with saving the text and rich text
setting, we'd request that the text be converted back into HTML for
display.

However, when getting the response from the server, we'd ignore that we
just forced rich text, and would set the rich text flag to that of the
text that was saved. So, the rich text coming in would be marked as
plain text. This caused the client to escape the HTML.

Now the rich text flag reflects the text coming from the server and not
the text being sent to the server.

Unit tests have been updated to ensure that the resulting flag is
correct in all cases.

Testing Done:
Unit tests pass.

Wrote the string `' " < > &` and sent as Markdown on a reply.
Saw it came back correctly from the server.

Then wrote the same string and sent as plain text. It no longer
appeared escaped.

Bugs closed: 3721

Reviewed at https://reviews.reviewboard.org/r/6802/