Fix scrolling and controls in diff fragments
Previously, when scrolling into an expanded diff fragment from below,
the diff collapse button would disappear off the screen if the diff
fragment took up the entire height of the browser. This was due to the
`outerHeight` method on the button being treated as a value and not a
function.

Fix an erroneous call to `_isMouseOverContainer` which should be
`_isMouseOverTargetContainer`. This fixes a bug where the controls
would not hide sometimes after an expand or collapse finished and the
mouse wasn't over the expanded container.