Update the docs database fixtures based on the latest schema.
The docs database fixtures haven't been updated in a while. This just
syncs up the generated fixtures with what exists in a newly-built
docs.db file. It'll help in keeping future changes to the fixtures a bit
smaller when up for review.

Testing Done:
Built the docs from the old fixtures and generated new ones. Deleted docs.db
and the build directory and built again, causing it to be regenerated from the
new fixtures. No problems.

Reviewed at https://reviews.reviewboard.org/r/6836/