Fix the image review UI comment serialization.
The generator comment serialization didn't work for the dummy review UI, so I
changed it to try to convert to a list. Unfortunately this broke the image
review UI, which serializes comments to a dictionary form.

I've changed it so that the basic review UI returns a list rather than an
iterator, and removed the explicit cast.

Testing done:
- Successfully opened up the image review UI with a comment present.
- Ran unit tests.

Fixes bug 3747.

Reviewed at https://reviews.reviewboard.org/r/6852/