Clicking username in nav bar now links to user profile instead of user preferences.
When you click on your avatar and/or username in the navigation bar, it takes you to the user preferences page. There is already a link to the Preferences page in the user drop down menu, but there isn't a link to your own user page anywhere. Your avatar and username in the navigation bar is now changed to link to your own profile page.

Testing Done:
Clicked my avatar/username from a variety of different places on the site. They all successfully linked to my own profile page.

Reviewed at https://reviews.reviewboard.org/r/6865/