Fix aggressive error logging for FileAttachment thumbnail and icon_url
With drag and drop inline images, we've introduced the ability to have a file
attachment with a null file field. In this case, we're expecting
`mimetype_handler` to be None (as there is no file). `thumbnail` and `icon_url`
can both be accessed when the file attachment does not yet have a file, and we
don't want to log an error in this case. So, just return None.

Testing Done:
All tests pass. There is no error message being logged to the console when
accessing `thumbnail` or `icon_url` properties on a file attachment with no
file.

Reviewed at https://reviews.reviewboard.org/r/6706/