Fix openIssue checkbox set to default when reloading the page with explicitly set value
In the "View Diff" page, if a comment was created without an issue opened, the
first time that it would be reopened after loading the page, the "Open an
Issue" check-box would be checked. The issue here was that we were not plumbing
the data through correctly, and the "loaded" state of the comment was being
used incorrectly to determine whether we could trust the initial data in the
page.

Testing Done:
Added a unit test, which passes along with all other tests.

Tested with a commentEditor which contains a comment whose `loaded` attribute
and `issueOpened` attribute are false and the default setting is true.
Previously, the editor's `openIssue` attribute is true. After, the attribute
will set to user's setting with the patch.

Test for some cases, all passed.

Bugs closed: 3580

Reviewed at https://reviews.reviewboard.org/r/6867/