Take the review banner into consideration in diff anchor offsets.
A recent change reduced the offset from the top of the page to the
selected anchor in the diff viewer, in order to show more of the diff.
This ended up causing the code to be covered up when showing the review
draft banner.

We now factor in the banner's height when choosing the offset from the
top of the page.

Testing Done:
Scrolled through the anchors with the keyboard shortcuts. Saw the small
gap between the top of the page and the anchor.

Added a comment and saw the draft banner appear. Scrolled through some
more, and saw the offset took the banner into consideration.

Reviewed at https://reviews.reviewboard.org/r/6912/