Don't break the URL in the diff viewer when reloading with an anchor.
When loading the diff viewer with an anchor, the "#" in the URL would be
stripped away, resulting in an incorrect, often broken URL. This was due
to `Backbone.history.start()` trying to convert it into a
pushState-compatible URL, with the assumption that it was supposed to be
a page in the URL.

We now save the hash, unset it, and then re-set it after
`Backbone.history.start()`.

Testing Done:
Navigated to an anchor and reloaded the page. Saw that the URL remained
correct, instead of turning into a broken URL.

Reviewed at https://reviews.reviewboard.org/r/6913/