Improve the admin UI for repositories and users.
The repository admin UI now allows for searching by repository names,
paths, and repository types. It also sorts alphabetically by name, and
allows the sort order to be changed.

The user admin UI embeds the Profile and LocalSiteProfiles associated
with the user. This makes it easier to locate a user and reset counts or
otherwise update information.

Testing Done:
Tested all the new capabilities in the list views and the inlines in the
User form.

Reviewed at https://reviews.reviewboard.org/r/6914/