Make the summary line for issue summaries contain rendered Markdown, instead of raw Markdown text.
This is the first step in redesigning the UI for the issue summary table (for
the end goal of speeding up the process of addressing issues).

Diff comments with Markdown text will be shown in the issue summary table
without their Markdown syntax. Example: A comment that says "This text is
**emphasized**" will be shown in the issue summary table as "This text is
emphasized". The  rendered Markdown text will be shown in the detail view for
the comment.

Testing Done:
Added comments with various Markdown text (bold, italic, fenced code block)
onto a review request. Formatting looked correct.

Reviewed at https://reviews.reviewboard.org/r/6870/