Make double-clicking on a row in a diff create a comment.
Users of GitHub or Bugzilla Splinter might be more used to creating
comments on diffs by double-clicking on the line of code they want
to comment on. This patch adds that capability.

Testing Done:
In a diff, I double-clicked on both sides of a diff, both within
syntax-highlighted regions, and outside of syntax-highlighted regions.
The comment editor dialog opened in all cases, as expected.

Reviewed at https://reviews.reviewboard.org/r/6881/