The repositoryCommitModel tests now check UTC fields
The tests in the `repositoryCommitModel` tests were parsing a UTC date
and expecting the the resulting year and date in the local timezone to
be the same as the year and date in UTC. As a result, this test fails in
certain timezones.

The test has been changed to check the UTC fields instead of the local
fields.

Testing Done:
Changed my timezone to UTC+8 and ran the JS tests. The test failed
prior to applying this patch.

Reviewed at https://reviews.reviewboard.org/r/6928/