Fix some PEP-257 issues in reviewboard/accounts/.
If we're eventually going to integrate the PEP-257 tool into Review Bot (which
I think we should), we should get our docstring house in order first. This
change goes through reviewboard/accounts and fixes most of the issues that
aren't just a missing docstring.

The one "error" that I chose to disable for this is D203, which wants a blank
line before class docstrings. I didn't see any justification for this in the
actual PEP-257 document, and I think it's ugly. There are also a couple cases
that I chose to ignore where our test docstrings which wrap didn't follow the
usual rules. There are a ton of things that have missing docstrings, but those
will require more work to clean up.

Testing done:
- Ran unit tests
- Ran pep257 --ignore=D203 reviewboard/accounts

Reviewed at https://reviews.reviewboard.org/r/6929/